Rename class params to follow naming standards
This changes and introduces class parameter names so that they follow
the naming convention of `thing_property`, e.g. `service_ensure`,
`package_name` or `package_ensure`. The updated parameter names are thus
used throughout the module.

It also moves all default values to the `kibana4::params` class,
including the dowload URL parameter which now defaults to `undef`. If
not specified in `kibana4::package_download_url` the proper download URL
is constructed using the Elasticsearch download site URL and
`package_name` and `package_ensure`.

The parameter documentation is updated in `init.pp` and in the README
file and the spec tests are updated with the new parameter names.