Run kibana service as kibana4 user by default
This sets the kibana4 user and group to "kibana4" by default,
independently of managing the user and group with this module.

The rationale is two-fold: 1) many Puppet sites manage system and user
accounts by means of custom user account modules, via LDAP/AD, or any
other means. Thus running the service as a user other than root should
be possible regardless of whether this module manages the user and
group.

2) It is better to not run the service as root:root in the first place
so let the module assume that a "kibana4" account already exists and use
it OR the administrator has specified a valid user/group via the
`kibana4::user` and `kibana4::group` parameters OR lets the module
manage the user and group account.